Revert "plat/rockchip: Use new bl31_params_parse_helper()"
This reverts commit 3e02c7436cf40fb7f7eb4d3038b7fc1ed1eeaa5f.

This is a hack to get arm-trusted-firmware to play nicely with
U-BOOT on the rockpi4b.  This reverts a change made to how BL31
receives the entry points for BL32/BL33.  This change reverts
back to a flat structure, whereas prior it was a linked list.

This change should be reverted once the proper fix is put in
place in U-BOOT.

BUG: 137954390
Test: local build and boot on the rockpi4b
Change-Id: Idc147a024261c01050bbfbbb8313bd8b4a1f7a48